dashboard corrected api name

+++ src/main/java/com/aristowebapi/dto/DashBoardCurrentMonthChart.java
@@ -2,12 +2,12 @@ package com.aristowebapi.dto;
 
 public interface DashBoardCurrentMonthChart {
 	
-	public long getSaleval();
-	public long getTargetval();
-	public long getLysval();
+	public double getSaleval();
+	public double getTargetval();
+	public double getLysval();
 	public double getAch();
 	public double getGth();
-	public long getSurdef();
+	public double getSurdef();
 	
 
 }
